Revert "Fix error with respect to EXPECTed WARNINGs/ERRRORs/..."
This reverts commit 2565678f23472e380e692ce2bbffa5d90309dda9.

The fix was not accepted by UVVM, with this comment message:
Hi Benny,
Passing an EXPECT to set_alert_attention() is not allowed, and you
should get a TB_WARNING in the transcript stating that only parameters
INGORE and REGARD are allowed (see UVVM Util QR section 1.3).

The purpose of the set_alert_attention() is to turn on/off alert
checking, while intentionally causing alerts, without knowing the exact
number of alerts you will receive.

When knowing the number of expected alerts, you should use the
increment_expeced_alerts(). The alerts will then be compared with the
specified stop_limit, be reported, and result in a simulation success or
failure.

We have created an internal issue to change from TB_WARNING to TB_ERROR
when passing EXPECT to set_alert_attention(), as this is an error in the
test bench and should not result in a warning.

Br.
Marius