Move bitvis_vip_wishbone to the UVVM repository.
The bitvis_vip_wishbone was in separate UVVM_Community repository,
however it needed to be copied to the UVVM repository to be compiled anyway.

There were also more problems with such approach.
Because it is not in the UVVM repository by default, the GHDL compile-uvvm.sh script does not support it, so it needed manual modifications.
All this is pure waste of time, putting bitvis_vip_wishbone into separate repository solved 0 problems, and created at least 2.